pyproject.toml
[tool.poetry]
name = "data-engineering-amazon-sales"
version = "0.1.0"
description = "ELT Pipeline or amazon sales"
authors = ["umairkarel <umair.karel@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "data_engineering_amazon_sales"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
snowflake = "^0.8.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
snowflake-connector-python = {extras = ["pandas"], version = "^3.10.1"}

[tool.pylint.messages_control]
disable="E0611, E0401, W0718, W0622, W0404"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
